Added SMAP SMEP checks in lookup_pte_as.

+++ includes/xstd/flat_map.hpp
@@ -281,7 +281,7 @@ namespace xstd
 
 				// Interpolate and get the mid point.
 				//
-				double interp = double( hash - begin->hash ) / double( std::prev( end )->hash - begin->hash );
+				float interp = float( hash - begin->hash ) / float( std::prev( end )->hash - begin->hash );
 				It mid = begin + std::clamp<size_t>( size_t( interp * ( count - 1 ) ), 1, count - 1 );
 
 				// If we found a matching entry:
